finish HW-4

+++ controller/index.js
@@ -1,17 +1,36 @@
 const service = require("../servisce");
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
+
+const SECRET_KEY = process.env.SECRET_KEY;
 
 const get = async (req, res, next) => {
+  const { id } = req.user;
+  const { page, limit, favorite } = req.query;
+  const skip = parseInt(page) > 1 ? (page - 1) * limit : 0;
   try {
-    const contacts = await service.getAllContacts();
-    return res.status(200).json(contacts);
+    const contacts = await service.getAllContacts({
+      owner: id,
+      skip,
+      limit,
+      favorite,
+    });
+    return res
+      .status(200)
+      .json({ data: contacts, amount: contacts.length, page, limit });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
 };
 
 const getById = async (req, res, next) => {
+  const { id } = req.user;
   try {
-    const contact = await service.getByIdContact(req.params.id);
+    const contact = await service.getByIdContact({
+      id: req.params.id,
+      owner: id,
+    });
     if (contact) {
       res.status(200).json(contact);
       return;
@@ -23,8 +42,9 @@ const getById = async (req, res, next) => {
 };
 
 const create = async (req, res, next) => {
+  const { id } = req.user;
   try {
-    const newContact = await service.createContact(req.body);
+    const newContact = await service.createContact({ ...req.body, owner: id });
     res.status(201).json(newContact);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -32,8 +52,9 @@ const create = async (req, res, next) => {
 };
 
 const remove = async (req, res, next) => {
+  const { id } = req.user;
   try {
-    const data = await service.removeContact(req.params.id);
+    const data = await service.removeContact({ id: req.params.id, owner: id });
     if (data) {
       res.status(200).json({ message: "contact deleted" });
       return;
@@ -45,6 +66,7 @@ const remove = async (req, res, next) => {
 };
 
 const update = async (req, res, next) => {
+  const owner = req.user.id;
   const { id } = req.params;
   const body = req.body;
 
@@ -54,8 +76,8 @@ const update = async (req, res, next) => {
       return;
     }
 
-    const putContact = await service.updateContact(id, body);
-    res.status(200).json({ ...putContact["_doc"], ...body });
+    const putContact = await service.updateContact({ id, body, owner });
+    res.status(200).json(putContact);
   } catch (error) {
     if (error.message.includes("Cast to ObjectId failed for value")) {
       res.status(400).json({ message: `Not found this id: ${id}` });
@@ -67,6 +89,7 @@ const update = async (req, res, next) => {
 };
 
 const favorite = async (req, res) => {
+  const owner = req.user.id;
   const { id } = req.params;
   const body = req.body;
 
@@ -76,11 +99,130 @@ const favorite = async (req, res) => {
       return;
     }
 
-    const data = await service.updateStatusContact(id, body);
-    res.status(200).json({ ...data["_doc"], ...body });
+    const data = await service.updateStatusContact({ id, body, owner });
+    res.status(200).json(data);
   } catch (error) {
     res.status(404).json({ message: "Not found" });
   }
 };
 
-module.exports = { get, getById, create, remove, update, favorite };
+/* ---------------------------------------------------------
+---------------------------/users/------------------------ 
+-----------------------------------------------------------*/
+
+/*============================ SIGNUP==================== */
+const signup = async (req, res) => {
+  const { email, password } = req.body;
+  try {
+    /* Перевіряємо є такий користувач у БД */
+    const user = await service.validateEmail(email);
+    if (user) {
+      res.status(409).json({ message: "Email in use" });
+      return;
+    }
+
+    /* Шифруємо пароль */
+    const hashPassword = await bcrypt.hash(password, 10);
+    /* Створюємо нового користувача */
+    const result = await service.createUser({ email, password: hashPassword });
+
+    /* Відправляємо відповідь */
+    res.status(201).json({
+      user: {
+        email: result.email,
+        subscription: "starter",
+      },
+    });
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+/*======================= LOGIN====================== */
+const login = async (req, res) => {
+  const { email, password } = req.body;
+  try {
+    /* Проверяем есть ли такой пользователь в БД */
+    const user = await service.validateEmail(email);
+    if (!user) {
+      res
+        .status(401)
+        .json({ message: `User not found with this email ${email}` });
+      return;
+    }
+
+    /* Проверяем пароль сходиться */
+    const checkPassword = await bcrypt.compare(password, user.password);
+    if (!checkPassword) {
+      res.status(401).json({ message: "The password is wrong" });
+      return;
+    }
+
+    /* Создаем токен */
+    const payload = { id: user["_id"], subscription: user.subscription };
+    const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "24h" });
+    /* Обновляем токен в БД */
+    const result = await service.updateUserToken({ id: user["_id"], token });
+
+    /* Отправляем ответ пользователю */
+    res.status(200).json({
+      token: result.token,
+      user: {
+        email: result.email,
+        subscription: result.subscription,
+      },
+    });
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+/*=========================== LOGOUT================= */
+const logout = async (req, res) => {
+  const { id } = req.user;
+  try {
+    await service.updateUserToken({ id, token: "" });
+    res.status(204).json();
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+/*=========================== CURRENT================= */
+const current = async (req, res) => {
+  const { email, subscription } = req.user;
+  try {
+    res.status(200).json({ email, subscription });
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+/*=========================== CURRENT================= */
+const subscription = async (req, res) => {
+  const { id } = req.user;
+  const { subscription } = req.body;
+  try {
+    const {
+      _id,
+      email,
+      subscription: newSub,
+    } = await service.updateSubscription({ id, subscription });
+    res.status(200).json({ _id, email, subscription: newSub });
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+module.exports = {
+  get,
+  getById,
+  create,
+  remove,
+  update,
+  favorite,
+  signup,
+  login,
+  logout,
+  current,
+  subscription,
+};
